refactor: auth 공통 exception은 common 폴더로 이동
